Revamped Small Enemy Ai
I reorganized the fixed the enemy AI by adding a state machine and streamlining the actions the enemy ai and take at any given moment. I also added a weapon system that allows the enemy to attack the player.

Next up is to make the enemy spawner which will spawn all the enemies each level and help dictate how challenging they should be.