Enemy Spawner
Created an enemy Spawner that dynamically spawns enemies according to the number of patrol Routes and can be set to spawn enemies at certain spawn points even with no patrol route.
It also controls the strength of the enemies with a difficulty curve that increases the further you are from spawn.